+++ lifelines/tests/utils/test_utils.py
@@ -187,12 +187,9 @@ def test_qth_survival_times_with_multivariate_q():
         pd.DataFrame([[40, 28], [25, 15]], index=[0.2, 0.5], columns=["sf", "sf**2"]),
     )
     assert_frame_equal(
-        utils.qth_survival_times([0.2, 0.5], sf_multi_df["sf"]),
-        pd.DataFrame([40, 25], index=[0.2, 0.5], columns=["sf"]),
+        utils.qth_survival_times([0.2, 0.5], sf_multi_df["sf"]), pd.DataFrame([40, 25], index=[0.2, 0.5], columns=["sf"])
     )
-    assert_frame_equal(
-        utils.qth_survival_times(0.5, sf_multi_df), pd.DataFrame([[25, 15]], index=[0.5], columns=["sf", "sf**2"])
-    )
+    assert_frame_equal(utils.qth_survival_times(0.5, sf_multi_df), pd.DataFrame([[25, 15]], index=[0.5], columns=["sf", "sf**2"]))
     assert utils.qth_survival_times(0.5, sf_multi_df["sf"]) == 25
 
 
@@ -475,9 +472,7 @@ def test_survival_table_from_events_raises_value_error_if_too_early_births():
 class TestLongDataFrameUtils(object):
     @pytest.fixture
     def seed_df(self):
-        df = pd.DataFrame.from_records(
-            [{"id": 1, "var1": 0.1, "T": 10, "E": 1}, {"id": 2, "var1": 0.5, "T": 12, "E": 0}]
-        )
+        df = pd.DataFrame.from_records([{"id": 1, "var1": 0.1, "T": 10, "E": 1}, {"id": 2, "var1": 0.5, "T": 12, "E": 0}])
         return utils.to_long_format(df, "T")
 
     @pytest.fixture
@@ -537,22 +532,13 @@ class TestLongDataFrameUtils(object):
         seed_df = seed_df[seed_df["id"] == 1]
 
         new_value_at_time_0 = 1.0
-        old_value_at_time_0 = seed_df["var1"]._iloc[0]
+        old_value_at_time_0 = seed_df["var1"].iloc[0]
         cv = pd.DataFrame.from_records([{"id": 1, "t": 0, "var1": new_value_at_time_0, "var2": 2.0}])
 
         df = seed_df.pipe(utils.add_covariate_to_timeline, cv, "id", "t", "E", overwrite=False)
 
         expected = pd.DataFrame.from_records(
-            [
-                {
-                    "E": True,
-                    "id": 1,
-                    "stop": 10.0,
-                    "start": 0,
-                    "var1": new_value_at_time_0 + old_value_at_time_0,
-                    "var2": 2.0,
-                }
-            ]
+            [{"E": True, "id": 1, "stop": 10.0, "start": 0, "var1": new_value_at_time_0 + old_value_at_time_0, "var2": 2.0}]
         )
         assert_frame_equal(df, expected, check_like=True)
 
@@ -564,9 +550,7 @@ class TestLongDataFrameUtils(object):
 
         df = seed_df.pipe(utils.add_covariate_to_timeline, cv, "id", "t", "E", overwrite=True)
 
-        expected = pd.DataFrame.from_records(
-            [{"E": True, "id": 1, "stop": 10.0, "start": 0, "var1": new_value_at_time_0}]
-        )
+        expected = pd.DataFrame.from_records([{"E": True, "id": 1, "stop": 10.0, "start": 0, "var1": new_value_at_time_0}])
         assert_frame_equal(df, expected, check_like=True)
 
     def test_enum_flag(self, seed_df, cv1, cv2):
@@ -582,9 +566,7 @@ class TestLongDataFrameUtils(object):
             # Windows Numpy and Pandas sometimes have int32 or int64 as default dtype
             if os.name == "nt" and "int32" in str(e) and "int64" in str(e):
                 assert_series_equal(
-                    df["enum"].loc[idx],
-                    pd.Series(np.arange(1, n + 1), dtype=df["enum"].loc[idx].dtypes),
-                    check_names=False,
+                    df["enum"].loc[idx], pd.Series(np.arange(1, n + 1), dtype=df["enum"].loc[idx].dtypes), check_names=False
                 )
             else:
                 raise e
@@ -665,9 +647,7 @@ class TestLongDataFrameUtils(object):
 
     def test_cumulative_sum(self):
         seed_df = pd.DataFrame.from_records([{"id": 1, "start": 0, "stop": 5, "E": 1}])
-        cv = pd.DataFrame.from_records(
-            [{"id": 1, "t": 0, "var4": 1}, {"id": 1, "t": 1, "var4": 1}, {"id": 1, "t": 3, "var4": 1}]
-        )
+        cv = pd.DataFrame.from_records([{"id": 1, "t": 0, "var4": 1}, {"id": 1, "t": 1, "var4": 1}, {"id": 1, "t": 3, "var4": 1}])
 
         df = seed_df.pipe(utils.add_covariate_to_timeline, cv, "id", "t", "E", cumulative_sum=True)
         expected = pd.DataFrame.from_records(
@@ -700,9 +680,7 @@ class TestLongDataFrameUtils(object):
 
     def test_covariates_from_event_matrix_with_simple_addition(self):
 
-        base_df = pd.DataFrame(
-            [[1, 0, 5, 1], [2, 0, 4, 1], [3, 0, 8, 1], [4, 0, 4, 1]], columns=["id", "start", "stop", "e"]
-        )
+        base_df = pd.DataFrame([[1, 0, 5, 1], [2, 0, 4, 1], [3, 0, 8, 1], [4, 0, 4, 1]], columns=["id", "start", "stop", "e"])
 
         event_df = pd.DataFrame([[1, 1], [2, 2], [3, 3], [4, None]], columns=["id", "poison"])
         cv = utils.covariates_from_event_matrix(event_df, "id")
@@ -725,9 +703,7 @@ class TestLongDataFrameUtils(object):
 
     def test_covariates_from_event_matrix(self):
 
-        base_df = pd.DataFrame(
-            [[1, 0, 5, 1], [2, 0, 4, 1], [3, 0, 8, 1], [4, 0, 4, 1]], columns=["id", "start", "stop", "e"]
-        )
+        base_df = pd.DataFrame([[1, 0, 5, 1], [2, 0, 4, 1], [3, 0, 8, 1], [4, 0, 4, 1]], columns=["id", "start", "stop", "e"])
 
         event_df = pd.DataFrame(
             [[1, 1, None, 2], [2, None, 5, None], [3, 3, 3, 7]], columns=["id", "promotion", "movement", "raise"]
@@ -818,9 +794,7 @@ class TestLongDataFrameUtils(object):
         rossi = load_rossi()
         rossi["id"] = np.arange(rossi.shape[0])
 
-        long_rossi = utils.to_episodic_format(
-            rossi, duration_col="week", event_col="arrest", id_col="id", time_gaps=1000.0
-        )
+        long_rossi = utils.to_episodic_format(rossi, duration_col="week", event_col="arrest", id_col="id", time_gaps=1000.0)
 
         # using astype(int) would fail on Windows because int32 and int64 are used as dtype
         long_rossi["week"] = long_rossi["stop"].astype(rossi["week"].dtype)
@@ -990,11 +964,7 @@ class TestSklearnAdapter:
 
         base_model = sklearn_adapter(WeibullAFTFitter, event_col="E")
 
-        grid_params = {
-            "penalizer": 10.0 ** np.arange(-2, 3),
-            "l1_ratio": [0.05, 0.5, 0.95],
-            "model_ancillary": [True, False],
-        }
+        grid_params = {"penalizer": 10.0 ** np.arange(-2, 3), "l1_ratio": [0.05, 0.5, 0.95], "model_ancillary": [True, False]}
         # note the n_jobs
         clf = GridSearchCV(base_model(), grid_params, cv=4, n_jobs=-1)
         clf.fit(X, Y)